+++ ocdskit/cli/commands/package_releases.py
@@ -1,24 +1,20 @@
-from collections import OrderedDict
-
 from .base import BaseCommand
+from ocdskit.combine import package_releases
 
 
 class Command(BaseCommand):
     name = 'package-releases'
-    help = 'reads releases from standard input, and prints one release package ' \
-           '(you will need to edit the package metadata)'
+    help = 'reads releases from standard input, and prints one release package'
 
     def add_arguments(self):
         self.add_argument('extension', help='add this extension to the package', nargs='*')
 
-    def handle(self):
-        releases = [self.json_loads(line) for line in self.buffer()]
+        self.add_package_arguments('release')
 
-        self.print(OrderedDict([
-            ('uri', 'http://example.com'),
-            ('publisher', {'name': ''}),
-            ('publishedDate', '9999-01-01T00:00:00Z'),
-            ('version', '1.1'),
-            ('extensions', self.args.extension),
-            ('releases', releases),
-        ]))
+    def handle(self):
+        kwargs = self.parse_package_arguments()
+        kwargs['extensions'] = self.args.extension
+
+        output = package_releases(self.buffer(), **kwargs)
+
+        self.print(output)