+++ ocdskit/cli/commands/upgrade.py
@@ -1,6 +1,7 @@
 from ocdskit import upgrade
 from ocdskit.cli.commands.base import BaseCommand
 from ocdskit.exceptions import CommandError
+from ocdskit.util import json_loads
 
 
 class Command(BaseCommand):
@@ -25,6 +26,6 @@ class Command(BaseCommand):
             raise CommandError('{}grade from {} is not supported'.format(direction, versions.replace(':', ' to ')))
 
         for line in self.buffer():
-            data = self.json_loads(line)
+            data = json_loads(line)
             upgrade_method(data)
             self.print(data)