+++ ocdskit/exceptions.py
@@ -4,3 +4,12 @@ class OCDSKitError(Exception):
 
 class CommandError(OCDSKitError):
     """Errors from within this package's CLI"""
+
+
+class InconsistentVersionError(OCDSKitError):
+    """Raised if the versions are inconsistent across packages to compile"""
+
+    def __init__(self, message, earlier_version=None, current_version=None):
+        self.earlier_version = earlier_version
+        self.current_version = current_version
+        super().__init__(message)